Moved from sql2o to JDBI
