+++ pom.xml
@@ -49,15 +49,6 @@
                         <goals>
                             <goal>shade</goal>
                         </goals>
-                        <configuration>
-                            <artifactSet>
-                                <includes>
-                                    <include>org.sql2o:sql2o</include>
-                                    <include>org.xerial:sqlite-jdbc</include>
-                                    <include>mysql:mysql-connector-java</include>
-                                </includes>
-                            </artifactSet>
-                        </configuration>
                     </execution>
                 </executions>
             </plugin>
@@ -84,9 +75,9 @@
         </dependency>
 
         <dependency>
-            <groupId>org.sql2o</groupId>
-            <artifactId>sql2o</artifactId>
-            <version>1.6.0</version>
+            <groupId>org.jdbi</groupId>
+            <artifactId>jdbi3-core</artifactId>
+            <version>3.32.0</version>
         </dependency>
 
         <dependency>
